型ごとにファイルを分離
