+++ Core/BGMBase.cs
@@ -2,44 +2,6 @@
 
 namespace MifuminSoft.funyan.Core
 {
-    public enum BGMMode
-    {
-        BGMM_NONE,
-        BGMM_DEFAULT,
-        BGMM_USER,
-    };
-
-    public enum BGMNumber
-    {
-        BGMN_SIRENT,
-        BGMN_TITLE,
-        BGMN_GAMEFUNYA,
-        BGMN_GAMEBANANA,
-        BGMN_GAMENEEDLE,
-        BGMN_GAMEGEASPRIN,
-        BGMN_GAMEWIND,
-        BGMN_GAMEICE,
-        BGMN_GAMEFIRE,
-        BGMN_GAMEMRFRAME,
-        BGMN_GAMEEELPITCHER,
-        BGMN_CLEAR,
-        BGMN_ENDING,
-        BGMN_MISS,
-        BGMN_GAMEOVER,
-        BGMN_EXPLAIN,
-        BGMN_SIZE,              // この番号のBGMは存在しない。配列のサイズを決めるため使う
-    };
-
-    public enum MENumber
-    {
-        MEN_SLEEP,
-        MEN_AWAKE,
-        MEN_GEASPRIN,
-        MEN_BANANADISTANCE,
-        MEN_BANANAPOSITION,
-        MEN_SIZE,               // この番号のBGMは存在しない。配列のサイズを決めるため使う
-    };
-
     public class CBGMBase : IDisposable
     {
         protected BGMNumber m_PlayNo;
