+++ Core/BGMMode.cs
@@ -1,0 +1,9 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum BGMMode
+    {
+        BGMM_NONE,
+        BGMM_DEFAULT,
+        BGMM_USER,
+    };
+}
