+++ Core/BGMNumber.cs
@@ -1,0 +1,23 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum BGMNumber
+    {
+        BGMN_SIRENT,
+        BGMN_TITLE,
+        BGMN_GAMEFUNYA,
+        BGMN_GAMEBANANA,
+        BGMN_GAMENEEDLE,
+        BGMN_GAMEGEASPRIN,
+        BGMN_GAMEWIND,
+        BGMN_GAMEICE,
+        BGMN_GAMEFIRE,
+        BGMN_GAMEMRFRAME,
+        BGMN_GAMEEELPITCHER,
+        BGMN_CLEAR,
+        BGMN_ENDING,
+        BGMN_MISS,
+        BGMN_GAMEOVER,
+        BGMN_EXPLAIN,
+        BGMN_SIZE,              // この番号のBGMは存在しない。配列のサイズを決めるため使う
+    };
+}
