+++ Core/CT.cs
@@ -1,0 +1,25 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum CT : uint
+    {
+        CT_TITL = 0x4C544954,
+        CT_AUTH = 0x48545541,
+        CT_DESC = 0x43534544,
+        CT_ENDM = 0x4D444E45,
+        CT_STGC = 0x43475453,
+        CT_REST = 0x54534552,
+        CT_PNLT = 0x544C4E50,
+        CT_HITS = 0x53544948,
+        CT_TL00 = 0x30304C54,
+        CT_M000 = 0x3030304D,
+        CT_M100 = 0x3030314D,
+        CT_M200 = 0x3030324D,
+        CT_MCF0 = 0x3046434D,
+        CT_MCD0 = 0x3044434D,
+
+        CT_RPLY = 0x594C5052,
+        CT_STGN = 0x4E475453,
+        CT_GRVT = 0x54565247,
+        CT_HYPR = 0x52505948,
+    }
+}
