+++ Core/F3KEY.cs
@@ -1,0 +1,23 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum F3KEY
+    {
+        F3KEY_EXIT,
+        F3KEY_PAUSE,
+        F3KEY_UP,
+        F3KEY_LEFT,
+        F3KEY_RIGHT,
+        F3KEY_DOWN,
+        F3KEY_JUMP,
+        F3KEY_ATTACK,
+        F3KEY_SMILE,
+        F3KEY_FPS,
+        F3KEY_BGMNONE,
+        F3KEY_BGMDEF,
+        F3KEY_BGMUSER,
+        F3KEY_CAPTURE,
+        F3KEY_RECORD,
+
+        F3KEY_BUFSIZE
+    }
+}
