+++ Core/Input.cs
@@ -2,34 +2,6 @@
 
 namespace MifuminSoft.funyan.Core
 {
-    public enum F3KEY
-    {
-        F3KEY_EXIT,
-        F3KEY_PAUSE,
-        F3KEY_UP,
-        F3KEY_LEFT,
-        F3KEY_RIGHT,
-        F3KEY_DOWN,
-        F3KEY_JUMP,
-        F3KEY_ATTACK,
-        F3KEY_SMILE,
-        F3KEY_FPS,
-        F3KEY_BGMNONE,
-        F3KEY_BGMDEF,
-        F3KEY_BGMUSER,
-        F3KEY_CAPTURE,
-        F3KEY_RECORD,
-
-        F3KEY_BUFSIZE
-    }
-
-    public struct tagButtonState
-    {
-        public int Button;     // 実際のボタン
-        public bool Pressed;   // 今押されているか
-        public bool Pushed;        // 今押されたか
-        public bool Released;   // 今離されたか
-    }
 
     /// <summary>
     /// Cf3Input
