+++ Core/MENumber.cs
@@ -1,0 +1,12 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum MENumber
+    {
+        MEN_SLEEP,
+        MEN_AWAKE,
+        MEN_GEASPRIN,
+        MEN_BANANADISTANCE,
+        MEN_BANANAPOSITION,
+        MEN_SIZE,               // この番号のBGMは存在しない。配列のサイズを決めるため使う
+    };
+}
