+++ Core/MapObjectBase.cs
@@ -3,30 +3,6 @@ using System.Collections.Generic;
 
 namespace MifuminSoft.funyan.Core
 {
-    public enum f3MapObjectDirection
-    {
-        DIR_FRONT = 0,
-        DIR_LEFT = 1,
-        DIR_RIGHT = 2,
-    }
-
-    public enum f3MapObjectType
-    {
-        //	MOT_UNKNOWN,
-
-        MOT_FUNYA,
-        MOT_BANANA,
-        MOT_NEEDLE,
-        MOT_GEASPRIN,
-        MOT_EELPITCHER,
-        MOT_ICE,
-        MOT_ICESOURCE,
-        MOT_FIRE,
-        MOT_EFFECT,
-
-        //	MOT_COUNT
-    }
-
     public abstract class Cf3MapObjectBase : IDisposable
     {
         private bool m_bValid;
