+++ Core/RID.cs
@@ -1,0 +1,18 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum RID
+    {
+        RID_MAIN,
+        RID_MAINICY,
+        RID_NEEDLE,
+        RID_GEASPRIN,
+        RID_EELPITCHER,
+        RID_ICEFIRE,
+        RID_MRFRAME,
+        RID_EFFECT,
+        RID_NAVI,
+        RID_HIT,
+
+        RID_END
+    }
+}
