+++ Core/f3JumpFunction.cs
@@ -1,0 +1,8 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public struct f3JumpFunction
+    {
+        public float DY;
+        public float Power;
+    };
+}
