+++ Core/f3MainCharaState.cs
@@ -1,0 +1,20 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum f3MainCharaState
+    {
+        STANDING,
+        RUNNING,
+        WALKING,
+        CHARGING,
+        JUMPING,
+        BREATHEIN,
+        BREATHEOUT,
+        TIRED,
+        DAMAGED,
+        FROZEN,
+        DEAD,
+        SMILING,
+        SLEEPING,
+        BLINKING,
+    }
+}
