+++ Core/f3MapObjectDirection.cs
@@ -1,0 +1,9 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum f3MapObjectDirection
+    {
+        DIR_FRONT = 0,
+        DIR_LEFT = 1,
+        DIR_RIGHT = 2,
+    }
+}
