+++ Core/f3MapObjectType.cs
@@ -1,0 +1,19 @@
+﻿namespace MifuminSoft.funyan.Core
+{
+    public enum f3MapObjectType
+    {
+        //	MOT_UNKNOWN,
+
+        MOT_FUNYA,
+        MOT_BANANA,
+        MOT_NEEDLE,
+        MOT_GEASPRIN,
+        MOT_EELPITCHER,
+        MOT_ICE,
+        MOT_ICESOURCE,
+        MOT_FIRE,
+        MOT_EFFECT,
+
+        //	MOT_COUNT
+    }
+}
